requirements.txt
six>=1.9.0
pylint>=1.4.4
pyensembl>=1.5.0
varcode
pandas>=0.23.0
nose>=1.3.3
progressbar2
pandas
